Move two helper functions out of `collection.rs`
